feat(placeholder,uix): add PlaceholderAPI integration and optimize rendering performance
PlaceholderAPI integration:
- Add PlaceholderService for periodic placeholder resolution and push
- Add ScreenPlaceholderConfig for per-screen placeholder and interval config
- Add PlaceholderStore for client-side placeholder value management
- Update UiConfigurations to extract placeholder configs from screen YAML
- Update huds/example.yml with placeholder usage examples

UIx rendering optimizations:
- Extract ColorUtil to deduplicate parseColor across all renderers
- Add ITextHolder interface for text-carrying widgets
- Add hasHoveredComponents() early exit in TooltipRenderer
- Reuse scratchSet in EventDispatcher to avoid per-frame allocation
- Optimize LayoutCache with Long key instead of CacheKey data class
- Refactor LayoutEngine resolveAnchoredPosition to reduce duplication
- Cache ResourceLocation in ImageRenderer to avoid repeated parsing
- Clean up ImageWidgetImpl render (now handled by ImageRenderer)

Performance and stability:
- Add packet serialization caching in NetworkHandler (invalidateCache)
- Cache RawAnimation instances in GenericReplacedEntity
- Add GeckoLib model cache check in GenericGeoRenderer to prevent crash
- Batch font icon rendering via MultiBufferSource in StringRenderOutputMixin
- Cache name component lookup in EntityMixin
- Remove redundant resource existence checks in RendererHandler
- Pre-build ScreenType cache in UiScreenStorage
- Use float for CharacterWrapper dimensions (was Int)
- Add rate limiting in LanternChannelMessageListener
- Cleanup player data on quit in PlayerListener
- Add connect/read timeout in TextureHandler HTTP requests
- Support encrypted pack reload in EncryptedPackLoader
- Fix TextUtil hex regex to support uppercase hex
- Update ResourceHandler rebuild to restore block and entity models
- Add disconnect cleanup and render pipeline diagnostics in LanternFabric
- Add block model and position packet parsing in NetworkParser
- Update Configurations to load blockModels.yml

+++ src/main/kotlin/org/lantern/internal/wrapper/key/CharacterWrapper.kt
@@ -4,9 +4,9 @@ import com.google.gson.JsonObject
 
 class CharacterWrapper(
     val resource: String,
-    val width: Int,
-    val height: Int,
-    val wide: Int
+    val width: Float,
+    val height: Float,
+    val wide: Float
 ) {
     companion object {
 
@@ -14,9 +14,9 @@ class CharacterWrapper(
             val texture = obj.get("texture").asString
             val wrapper = CharacterWrapper(
                 texture,
-                obj.get("width").asInt,
-                obj.get("height").asInt,
-                obj.get("wide").asInt
+                obj.get("width").asFloat,
+                obj.get("height").asFloat,
+                obj.get("wide").asFloat
             )
             return wrapper
         }
